Added live updating to the Build Switcher in the Pipeline Header
